Filter host names by optional SERVICE_NAME prefix.
Allows you to specify the SERVICE_NAME environment variables, which is a csv of prefixes for host entries to be included in the list of backends for haproxy.

The default value for SERVICE_NAME (declared in the Dockerfile) is empty, which will match anything -- so this shouldn't break existing configurations.